Changed some names and fixed some bugs

+++ DV1573---UD1448/Spells/SpellHandler.cpp
@@ -433,7 +433,7 @@ void SpellHandler::spellCollisionCheck()
 						if (m_onHitCallback != nullptr) {
 							m_onHitCallback();
 						}
-						k = m_nrSubSteps;
+						k = static_cast<size_t>(m_nrSubSteps);
 					}
 				}
 			}
